refactor custom classes
